Add test for downloading (localhost, single instance).
The bash test driver is improved to create a temporary working dir for
each test. An empty uftt.dat is created there for use by uftt instances,
and the test can create files and directories there. Everything will be
cleaned up by the test driver when the test is successfull.