Implement --notify-socket command line option.
This option, with a port number as argument, will open a TCP connection
to the specified port on localhost. When the connection is made the
socket will be closed.

This is usefull for scripting/testing, as you can wait for such a
connection and be sure uftt is in a state where it will accept
local connections from further uftt instances.