[net-asio] Silence last warning: the type of nlmsghdr.nlmsg_len is __u32, not int:
This also fixes a idiosyncrasy in the existing code: ilen and alen appear
to be the same type of quantity, but they are being treated slightly
differently. In particular 'alen' is checked first to be smaller than the
size of the pointer 'ifa', and then later on again the struct 'ifa',
without being used in-between. The first check seems to be unintentional,
since comparing against the size of a pointer makes little sense to me, so
I've removed that one and moved the latter check up a bit.